test helper moved to internal folder
