+++ pkg/config/directory/group_test.go
@@ -7,8 +7,8 @@ import (
 	"testing"
 
 	"github.com/google/uuid"
+	"github.com/ibm-verify/verify-sdk-go/internal/test_helper"
 	"github.com/ibm-verify/verify-sdk-go/pkg/auth"
-	"github.com/ibm-verify/verify-sdk-go/pkg/config/config_test"
 	"github.com/ibm-verify/verify-sdk-go/pkg/config/directory"
 	"gopkg.in/yaml.v3"
 
@@ -36,7 +36,7 @@ func (s *GroupTestSuite) SetupTest() {
 	logger.AddNewline = true
 
 	// load common config
-	tenant, clientID, clientSecret := config_test.LoadCommonConfig(s.T())
+	tenant, clientID, clientSecret := test_helper.LoadCommonConfig(s.T())
 
 	// get token
 	client := &auth.Client{
